@@ -47,13 +47,12 @@
   },
   "dependencies": {},
   "peerDependencies": {
+    "@carbon/react": "^1.9.0",
     "@openmrs/esm-framework": "next",
-    "carbon-icons": "7.x",
-    "dayjs": "1.x",
-    "react": "^16.14",
-    "react-dom": "^16.0",
+    "react": "^18.x",
+    "react-dom": "^18.x",
     "react-i18next": "11.x",
-    "react-router-dom": "^5.0.0"
+    "react-router-dom": "^6.3"
   },
   "devDependencies": {
     "@babel/core": "^7.14.3",
@@ -62,18 +61,15 @@
     "@babel/preset-react": "^7.13.13",
     "@babel/preset-typescript": "^7.13.0",
     "@babel/runtime": "^7.14.0",
-    "@openmrs/esm-framework": "next",
+    "@openmrs/esm-framework": "^4.0.1-pre.208",
     "@testing-library/dom": "^7.20.0",
     "@testing-library/jest-dom": "^5.11.0",
     "@testing-library/react": "^10.4.3",
     "@testing-library/user-event": "^12.0.11",
-    "@types/carbon-components-react": "^7.34.0",
-    "@types/carbon__icons-react": "^10.31.0",
-    "@types/jest": "^26.0.23",
-    "@types/react": "^16.14.5",
-    "@types/react-dom": "^16.9.12",
-    "@types/react-router-dom": "^5.1.7",
-    "@types/testing-library__jest-dom": "^5.14.1",
+    "@types/jest": "^28.1.6",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/testing-library__jest-dom": "^5.14.5",
     "@types/webpack-env": "^1.16.0",
     "@typescript-eslint/parser": "^5.14.0",
     "babel-eslint": "^10.1.0",
@@ -87,15 +83,12 @@
     "eslint-plugin-prettier": "^4.2.1",
     "husky": "^8.0.1",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^26.4.2",
-    "jest-cli": "^26.4.2",
-    "openmrs": "next",
+    "jest": "^28.1.3",
+    "jest-cli": "^28.1.3",
+    "jest-environment-jsdom": "^28.1.3",
+    "openmrs": "^4.0.1-pre.208",
     "prettier": "^2.3.0",
     "pretty-quick": "^3.1.0",
-    "react": "^16.14",
-    "react-dom": "^16.0.0",
-    "react-i18next": "^11.16.9",
-    "react-router-dom": "^5.0.0",
     "swc-loader": "^0.2.3",
     "swr": "^1.3.0",
     "typescript": "^4.7.3",
