@@ -1,5 +1,5 @@
 import React from "react";
-import { BrowserRouter, Route, Switch } from "react-router-dom";
+import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { appPath } from "./constant";
 const FormsPage = React.lazy(() => import("./forms-page"));
 const FormEntryWorkflow = React.lazy(() => import("./form-entry-workflow"));
@@ -8,10 +8,10 @@ const Root = () => {
   return (
     <main>
       <BrowserRouter basename={appPath}>
-        <Switch>
-          <Route exact path="/" children={<FormsPage />} />
-          <Route path="/:formUuid?" children={<FormEntryWorkflow />} />
-        </Switch>
+        <Routes>
+          <Route path="" element={<FormsPage />} />
+          <Route path="/:formUuid" element={<FormEntryWorkflow />} />
+        </Routes>
       </BrowserRouter>
     </main>
   );
