@@ -1,23 +1,23 @@
-import { Button } from "carbon-components-react";
+import { Button } from "@carbon/react";
 import React, { useContext } from "react";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import FormWorkflowContext from "../context/FormWorkflowContext";
 import FormReviewCard from "../form-review-card";
 import styles from "./styles.scss";
 
 const WorkflowReview = () => {
   const { patientUuids } = useContext(FormWorkflowContext);
-  const history = useHistory();
+  const navigate = useNavigate();
   return (
     <div className={styles.workspaceWrapper}>
       <div className={styles.workspace}>
         <div className={styles.leftPanel}>
           <h4>Review</h4>
           <div className={styles.navButtons}>
-            <Button kind="primary" onClick={() => history.push("/")}>
+            <Button kind="primary" onClick={() => navigate("/")}>
               Save & Close
             </Button>
-            <Button kind="tertiary" onClick={() => history.push("/")}>
+            <Button kind="tertiary" onClick={() => navigate("/")}>
               Cancel
             </Button>
           </div>
