@@ -2,7 +2,7 @@ import React, { useEffect, useMemo, useReducer } from "react";
 import reducer from "./FormWorkflowReducer";
 import { useParams, useLocation } from "react-router-dom";
 interface ParamTypes {
-  formUuid: string;
+  formUuid?: string;
 }
 
 const initialActions = {
