@@ -1,5 +1,5 @@
 import React from "react";
-import { Tile } from "carbon-components-react";
+import { Tile, Layer } from "@carbon/react";
 import styles from "./styles.scss";
 import { useLayoutType } from "@openmrs/esm-framework";
 import { EmptyDataIllustration } from "./EmptyDataIllustration";
@@ -15,13 +15,17 @@ const EmptyState: React.FC<EmptyStateProps> = ({
   const isTablet = useLayoutType() === "tablet";
 
   return (
-    <Tile light className={styles.tile}>
-      <div className={isTablet ? styles.tabletHeading : styles.desktopHeading}>
-        <h4>{headerTitle}</h4>
-      </div>
-      <EmptyDataIllustration />
-      <p className={styles.content}>{displayText}</p>
-    </Tile>
+    <Layer>
+      <Tile className={styles.tile}>
+        <div
+          className={isTablet ? styles.tabletHeading : styles.desktopHeading}
+        >
+          <h4>{headerTitle}</h4>
+        </div>
+        <EmptyDataIllustration />
+        <p className={styles.content}>{displayText}</p>
+      </Tile>
+    </Layer>
   );
 };
 
